+++ include/Condition.hpp
@@ -9,23 +9,23 @@ class Condition : NonCopyable {
  public:
   explicit Condition(MutexLock& mutex)
     : mutex_(mutex) {
-    LIMONP_CHECK(!pthread_cond_init(&pcond_, NULL));
+    CHECK(!pthread_cond_init(&pcond_, NULL));
   }
 
   ~Condition() {
-    LIMONP_CHECK(!pthread_cond_destroy(&pcond_));
+    CHECK(!pthread_cond_destroy(&pcond_));
   }
 
   void Wait() {
-    LIMONP_CHECK(!pthread_cond_wait(&pcond_, mutex_.GetPthreadMutex()));
+    CHECK(!pthread_cond_wait(&pcond_, mutex_.GetPthreadMutex()));
   }
 
   void Notify() {
-    LIMONP_CHECK(!pthread_cond_signal(&pcond_));
+    CHECK(!pthread_cond_signal(&pcond_));
   }
 
   void NotifyAll() {
-    LIMONP_CHECK(!pthread_cond_broadcast(&pcond_));
+    CHECK(!pthread_cond_broadcast(&pcond_));
   }
 
  private: