+++ include/Thread.hpp
@@ -1,7 +1,7 @@
 #ifndef LIMONP_THREAD_HPP
 #define LIMONP_THREAD_HPP
 
-#include "HandyMacro.hpp"
+#include "Logging.hpp"
 #include "NonCopyable.hpp"
 
 namespace limonp {
@@ -12,19 +12,19 @@ class IThread: NonCopyable {
   }
   virtual ~IThread() {
     if(isStarted && !isJoined) {
-      LIMONP_CHECK(!pthread_detach(thread_));
+      CHECK(!pthread_detach(thread_));
     }
   };
 
   virtual void Run() = 0;
   void Start() {
-    LIMONP_CHECK(!isStarted);
-    LIMONP_CHECK(!pthread_create(&thread_, NULL, Worker, this));
+    CHECK(!isStarted);
+    CHECK(!pthread_create(&thread_, NULL, Worker, this));
     isStarted = true;
   }
   void Join() {
-    LIMONP_CHECK(!isJoined);
-    LIMONP_CHECK(!pthread_join(thread_, NULL));
+    CHECK(!isJoined);
+    CHECK(!pthread_join(thread_, NULL));
     isJoined = true;
   }
  private: